frontend/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.0.1
aiohttp==3.9.3
pymongo==4.6.3
telegram==0.0.1
flask==3.0.3